+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Change Log
 
 - [trivial] Dim the switcher buttons in the inspector when the window is inactive.
 - [trivial] Improve the style of the pop-up buttons in the print panel.
+- [trivial] Adjust the layout of the regular expression syntax reference.
 
 
 ### Fixes